[interop] Moved PyTorch wrapper under operator module with other wrappers.
