--- src/pycsou/interop/__init__.py
@@ -1,2 +1,0 @@
-from _wrapper import *
-from pytorch import *
