+++ src/pycsou_tests/operator/interop/test_from_sciop.py
@@ -3,7 +3,7 @@ import itertools
 import pytest
 import scipy.sparse.linalg as spsl
 
-import pycsou.operator.interop.scipy as isp
+import pycsou.operator.interop.sciop as isp
 import pycsou.runtime as pycrt
 import pycsou.util.deps as pycd
 import pycsou.util.ptype as pyct
